created binary-search.py
apparently, there are some better alternatives to linear search. this is one of them. unfortunately, i did not explain the algo in the code itself, didnt have much time. anyways, lets compensate that here. so, basically, binary search in python was quite simple, the concept is that given an array of n items, we take the first item and the last item's indices, and we assign them the value of first and last respectively. then, we just find the average of those 2. for that, we use // which is called integral division, where we dont return a float value, because we dont want that. then, if that average value is equal to the number, we show the number. else, we see if the number we are looking for is greater or not than the number we are looking for. if it is, then we make this number the 'first' variable, and if it is not, then we make this number the 'last' value, and we repeat the process. the time complexity of this algorithm is O(log n)